requirements.txt
numpy
scipy
soundfile
pyloudnorm
pyyaml
